refactor: recoil 관련 파일 조정
